pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "scene-query"
version = "0.1.0"
description = "Natural language queries over 3D scenes (Gaussian Splats, NeRFs, Point Clouds)"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Raj Gandhi" }]

dependencies = [
    "open-clip-torch>=2.24.0",
    "segment-anything>=1.0",
    "faiss-cpu>=1.8.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.0.0",
    "torch>=2.3.0",
    "torchvision>=0.18.0",
    "numpy>=1.26.0",
    "msgpack>=1.0.8",
    "PyYAML>=6.0",
    "Pillow>=10.0.0",
    "scipy>=1.13.0",
    "tqdm>=4.66.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
# faiss-gpu is Linux-only; on Linux with CUDA, replace faiss-cpu in dependencies with faiss-gpu==1.7.2
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-benchmark>=4.0.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
    "types-PyYAML>=6.0",
    "types-tqdm>=4.66.0",
]

[project.scripts]
scene-query = "python.api.app:main"

[tool.hatch.build.targets.wheel]
packages = ["python"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "benchmark: performance benchmark tests",
    "integration: integration tests requiring model weights",
    "e2e: end-to-end tests requiring full stack",
]
